scripts/requirements.txt
requests
tqdm
PyYAML
pillow
numpy
opencv-python
exifread
imagehash